LayoutTests/touchadjustment/zoom-basic-expected.txt

Test basic cases of tap-to-zoom mechanics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS zoomableArea.top is 100
PASS zoomableArea.left is 100
PASS zoomableArea.width is 200
PASS zoomableArea.height is 100
PASS zoomableArea.top is 210
PASS zoomableArea.left is 110
PASS zoomableArea.width is 180
PASS zoomableArea.height is 80
PASS zoomableArea.top is 200
PASS zoomableArea.left is 100
PASS zoomableArea.width is 200
PASS zoomableArea.height is 100
PASS zoomableArea.top is 50
PASS zoomableArea.left is 50
PASS zoomableArea.width is 300
PASS zoomableArea.height is 300
PASS successfullyParsed is true

TEST COMPLETE
